pyproject.toml
[project]
name = "okxtradingbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=5.0.0",
    "ccxt>=4.5.37",
    "numpy>=2.2.6",
    "pandas>=2.3.3",
    "pandas-ta>=0.4.71b0",
    "plotly>=6.5.2",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=22.6",
    "requests>=2.32.5",
    "scikit-learn>=1.8.0",
    "sqlalchemy>=2.0.46",
    "streamlit>=1.54.0",
]
